requirements.txt

click==8.1.7
Flask==3.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
Flask-Mail==0.9.1
python-dotenv==1.0.0